Implement murmur hash 3.

+++ list/src/list.hh
@@ -109,7 +109,7 @@ public:
 //States
 //  hash
 //  [inherited from Object]
-  int hash() const;
+  unsigned long long hash() const;
 //  clone
 //  [inherited from Object]
 //memory management may be up to the user.
